backend/requirements.txt
websockets
openai
SpeechRecognition
PyAudio
pyttsx3
python-dotenv
aiofiles
AudioConverter